Fix click issue with topology context menu
